finished coding some baselines

+++ oem/losses.py
@@ -70,7 +70,7 @@ class FocalLoss(nn.Module):
 
     def forward(self, input, target):
         bce_loss = nn.functional.binary_cross_entropy_with_logits(
-            input, target, reduction="none"
+            input, target.float(), reduction="none"
         )
 
         pt = torch.exp(-bce_loss)
